+++ src/compareworkspace_p.h
@@ -16,6 +16,7 @@
 
 #include "core/EventBus.h"
 #include "core/compare/DifferenceEngine.h"
+#include "core/image/ImageRepository.h"
 #include "core/compare/Histogram.h"
 #include "core/image/ImageBuffer.h"
 #include "core/image/QtConvert.h"